Update deduplica.py
I optimized the script to automatically check if the myfile.db database exists every time any function is launched. If the database does not exist, the script will create it automatically, including the necessary table. This process takes place within the create_connection function, which now includes a call to ensure_database_exists.

I updated the script making the db_file parameter optional in all functions. If the user does not specify a path to the database file, the script will automatically use myfile.db as default. I made this change in all command definitions in the command line interface.